Components reorganization
