+++ packages/fe-container/src/components/layout-content/LayoutContent.tsx
@@ -1,8 +1,11 @@
 import React, {useCallback, useContext} from 'react'
 import {Layout} from 'antd'
+import {Plugin} from '@mia-platform/core'
+import {Route, Router, Switch} from 'react-router-dom'
 
 import {MenuOpenedContext} from '../../contexts/MenuOpened.context'
-import {LayoutCenter} from './layout-center/LayoutCenter'
+import {ConfigurationContext} from '../../contexts/Configuration.context'
+import {history} from '../../plugins/PluginsLoaderFacade'
 
 import './LayoutContent.less'
 
@@ -15,5 +18,35 @@ export const LayoutContent: React.FC = () => {
     <Layout className={isMenuOpened ? 'layout-container-overlay' : ''} onClick={closeSideMenu}>
       <LayoutCenter/>
     </Layout>
+  )
+}
+
+const LayoutCenter: React.FC = () => {
+  const configuration = useContext(ConfigurationContext)
+  const hasRoute = useCallback((plugin: Plugin) => plugin.pluginRoute, [])
+  const routerMapper = useCallback((plugin: Plugin) => (
+    <Route key={plugin.id} path={plugin.pluginRoute}>
+      <CenterPluginManager {...plugin}/>
+    </Route>
+  ), [])
+
+  return (
+    <Layout.Content data-testid="layout-content-overlay">
+      <Router history={history}>
+        <Switch>
+          {configuration.plugins?.filter(hasRoute).map(routerMapper)}
+        </Switch>
+      </Router>
+    </Layout.Content>
+  )
+}
+
+const CenterPluginManager: React.FC<Plugin> = (plugin) => {
+  return (
+    <>
+      {plugin.integrationMode === 'iframe' &&
+      <iframe className="layout-iframe" src={plugin.pluginUrl} title={plugin.id}/>}
+      {plugin.integrationMode === 'qiankun' && <div id={plugin.id}/>}
+    </>
   )
 }
