+++ packages/fe-container/src/components/top-bar/TopBar.tsx
@@ -1,12 +1,8 @@
-import React, {useContext, useState} from 'react'
+import React from 'react'
 import {FormattedMessage} from 'react-intl'
-import {Dropdown} from 'antd'
-
-import {MenuOpenedContext} from '../../contexts/MenuOpened.context'
-import {SideMenu} from '../side-menu/SideMenu'
-import {ConfigurationContext} from '../../contexts/Configuration.context'
 
 import './TopBar.less'
+import {BurgerIcon} from '../burger-icon/BurgerIcon'
 
 export const TopBar: React.FC = () => {
   return (
@@ -16,33 +12,5 @@ export const TopBar: React.FC = () => {
         <FormattedMessage id="topBarTitle"/>
       </span>
     </div>
-  )
-}
-
-const BurgerIcon: React.FC = () => {
-  const {isMenuOpened, setMenuOpened} = useContext(MenuOpenedContext)
-  const [isChecked, setChecked] = useState(false)
-
-  const manageToggle = () => {
-    setChecked(!isChecked)
-    setMenuOpened(!isMenuOpened)
-  }
-
-  const configuration = useContext(ConfigurationContext)
-
-  const sideMenu = <SideMenu plugins={configuration?.plugins}/>
-
-  return (
-    <Dropdown
-      onVisibleChange={manageToggle} overlay={sideMenu} overlayClassName="sideMenu_overlay"
-      trigger={['click']} visible={isMenuOpened}
-    >
-      <label htmlFor="check">
-        <input checked={isMenuOpened} data-testid="top-bar-side-menu-toggle" readOnly={true} type="checkbox"/>
-        <span/>
-        <span/>
-        <span/>
-      </label>
-    </Dropdown>
   )
 }
